Strip redundant identity casts on u8 fields/registers
After the u8 migration, many (X as i32) as u8 / X as u8 as i32 casts where X is
already a u8 field or register are pure identities. Remove them (targeting only
known-u8 operands; genuine i32-expression truncations are left intact). Call
sites now read as clean field access, e.g.
  r.value = engine.state.player_health;
  engine.state.player_health = engine.state.player_health - 1;
Verified byte-exact: 800-frame trace identical, 34 test bins green.

+++ src/game.rs
@@ -5744,7 +5744,7 @@ pub fn restore_status_sprite_template(engine: &mut Engine, r: &mut RoutineContex
 /// Spends one health point, returning carry set when health was already
 /// empty.
 pub fn consume_health_point(engine: &mut Engine, r: &mut RoutineContext) {
-    r.value = ((engine.state.player_health as i32) as u8);
+    r.value = engine.state.player_health;
     if (r.value == 0) {
         r.carry = 1;
         return;
@@ -5774,7 +5774,7 @@ pub fn subtract_health_points(engine: &mut Engine, r: &mut RoutineContext) {
 /// set when no magic was available.
 pub fn consume_magic_point(engine: &mut Engine, r: &mut RoutineContext) {
     let saved_index: i32 = ((r.index) as u8 as i32);
-    r.value = ((engine.state.player_magic as i32) as u8);
+    r.value = engine.state.player_magic;
     r.carry = 1;
     if (engine.state.player_magic != 0) {
         engine.state.player_magic = engine.state.player_magic - 1;
